Cex dex tests

Former-commit-id: 3b0e595dab4050b4c942e3a32c8f1c0fcacd5e16

+++ crates/brontes-classifier/src/test_utils.rs
@@ -163,7 +163,7 @@ pub async fn get_traces_with_meta(
     (traces, header, metadata)
 }
 
-fn get_coinbase_transfer(builder: Address, action: &Action) -> Option<u64> {
+fn get_coinbase_transfer(builder: Address, action: &Action) -> Option<u128> {
     match action {
         Action::Call(action) => {
             if action.to == builder {
